Move Short/SuffixedNino to domain package
